pyproject.toml
[project]
name = "wos-code-redeem"
version = "0.1.0"
description = "CLI to redeem gift codes via Century Game API"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31,<3",
  "fastapi>=0.112",
  "uvicorn>=0.30",
  "sqlalchemy>=2.0",
  "jinja2>=3.1",
  "feedparser>=6.0",
  "psycopg[binary]>=3.2",
  "passlib[bcrypt]>=1.7",
  "itsdangerous>=2.2",
  "python-multipart>=0.0.9",
  "alembic>=1.13",
  "PyJWT[crypto]>=2.9",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["wos_redeem"]
